Update Rust version to 1.74.1
We originally pinned to 1.63 because that was the last version before
Rust dropped support for older versions of `glibc` [in Rust 1.64][blog].
However, the relevant operating systems we were maintaining support for
at that time (CentOS/RHEL 6) have now been end-of-life for three full
years, so it seems reasonable to update to a more current version. Since
no other changes to the `glibc` support have happened in the interval,
we can safely update to the most current version of Rust.

As part of this, update the RHEL 7 Dockerfile setup for building on Rust
1.74, including using updated versions of gcc and binutils.

[blog]: https://blog.rust-lang.org/2022/08/01/Increasing-glibc-kernel-requirements.html

Co-authored-by: Rob Jackson <me@rwjblue.com>